requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
PyJWT[crypto]
cryptography
aiofiles
httpx
python-dotenv
requests
passlib[bcrypt]
email-validator

# LTI 1.3 specific dependencies
jwcrypto

# Development dependencies
pytest
pytest-asyncio
black
flake8